pasajeros-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
mysqlclient==2.2.0
email-validator==2.1.0.post1
requests==2.31.0
tenacity==8.2.3
prometheus-client==0.19.0
python-json-logger==2.0.7
pytest==7.4.3
aio-pika==9.3.0
httpx==0.25.2
pytest-asyncio==0.21.1
alembic==1.12.1
pybreaker==1.3.0
locust==2.24.0
redis==4.5.5
fastapi-cache2==0.2.1
fastapi-limiter==0.1.5
fastapi-health==0.4.0 